composite: move it under Xext/
