+++ Xext/composite/compalloc.c
