+++ Xext/composite/compint.h
