+++ Xext/composite/compoverlay.c
