+++ Xext/meson.build
@@ -56,3 +56,4 @@ subdir('dri3')
 subdir('panoramiX')
 subdir('present')
 subdir('xf86bigfont')
+subdir('composite')
